backend/requirement.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
google-genai